modified stuffs

+++ base/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render,redirect
 from .models import Room,Topic
 from django.contrib.auth.models import User
+from django.http import HttpResponse
 from django.contrib.auth import authenticate,login,logout
 from .forms import RoomForm
 from django.contrib.auth.decorators import login_required
@@ -13,6 +14,8 @@ rooms=[
     {'id':3,'name':'Nice work'},
 ]
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
     if request.method=="POST":
         username=request.POST.get('username')
         password=request.POST.get('password')
@@ -58,9 +61,12 @@ def createRoom(request):
     context={"form":form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='/login')
 def updateRoom(request,pk):
     room=Room.objects.get(id=pk)
     form=RoomForm(instance=room)
+    if request.user!=room.host:
+        return HttpResponse("You are not allowed here!")
     
     if request.method=="POST":
         form=RoomForm(request.POST,instance=room)
@@ -70,8 +76,11 @@ def updateRoom(request,pk):
     context={'form':form}
     return render(request,"base/room_form.html",context)
 
+@login_required(login_url='/login')
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+    if request.user!=room.host:
+        return HttpResponse("You are not allowed here!")
     if request.method=='POST':
         room.delete()
         return redirect('home')
